+++ management/internals/modules/agentnetwork/affectedpeers_hook.go
