+++ management/internals/modules/agentnetwork/catalog/catalog.go
