+++ management/internals/modules/agentnetwork/policyselect.go
@@ -7,7 +7,7 @@ import (
 	"sort"
 	"time"
 
-	"github.com/netbirdio/netbird/management/server/agentnetwork/types"
+	"github.com/netbirdio/netbird/management/internals/modules/agentnetwork/types"
 	"github.com/netbirdio/netbird/management/server/store"
 	"github.com/netbirdio/netbird/shared/management/status"
 )
