+++ management/internals/modules/agentnetwork/types/accesslog.go
