+++ management/internals/modules/agentnetwork/types/consumption.go
