+++ management/internals/modules/agentnetwork/types/consumption_test.go
