+++ management/internals/modules/agentnetwork/types/settings.go
